requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
open3d>=0.13.0
laspy>=2.0.0
networkx>=2.6
opencv-python>=4.5.0
matplotlib>=3.5.0
pyyaml>=6.0